requirements.txt
pandas
numpy
joblib
shap
google-generativeai
scikit-learn
